Update README.md
This is a filterable phonebook that outputs the total number of contacts stored on it, allows users to add new contacts, search through to find a contact by filter, and it also sorts the contacts alphabetically.

I built this using HTML, CSS and vanilla JavaScript. The UI was inspired by the UI of a Samsung phone.

In the next update, I intend to randomly assign styles to contacts on the list, and enable local storage of contacts list.